requirements.txt
matplotlib==3.10.1
numpy==2.2.4
Pillow==11.2.1
